refactor: pull up common code to `BaseIntegrationTest` and extract `OPAPathSelectorTest` to a separate class
