+++ src/test/java/com/styra/opa/springboot/OPAAuthorizationManagerTest.java
@@ -1,55 +1,19 @@
 package com.styra.opa.springboot;
 
-import com.fasterxml.jackson.databind.JsonNode;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.styra.opa.OPAClient;
-import com.styra.opa.springboot.autoconfigure.OPAAutoConfiguration;
 import com.styra.opa.springboot.autoconfigure.OPAProperties;
-import jakarta.servlet.http.HttpServletRequest;
-import org.junit.jupiter.api.AfterEach;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
-import org.mockito.MockitoAnnotations;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Import;
-import org.springframework.core.Ordered;
-import org.springframework.core.annotation.Order;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.authorization.AuthorizationDecision;
 import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.web.access.intercept.RequestAuthorizationContext;
-import org.springframework.security.web.authentication.WebAuthenticationDetails;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.BindMode;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.containers.output.Slf4jLogConsumer;
-import org.testcontainers.images.builder.ImageFromDockerfile;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
 
 import java.net.URI;
 import java.net.http.HttpClient;
 import java.net.http.HttpRequest;
 import java.net.http.HttpResponse;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
-import java.util.Set;
 import java.util.function.Supplier;
 
 import static com.styra.opa.springboot.OPAAuthorizationManager.ACTION;
@@ -77,91 +41,9 @@ import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.anyString;
-import static org.mockito.Mockito.doReturn;
-import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
-@SpringBootTest(classes = OPAAutoConfiguration.class)
-@Testcontainers
-class OPAAuthorizationManagerTest {
-
-    private static final int OPA_PORT = 8181;
-    /**
-     * Besides OPA default port ({@code opaPort}), OPA is exposed under {@code localhost:8282/customprefix} using
-     * Nginx.<br/>
-     * Nginx configuration is provided in {@code src/test/resources/nginx.conf} file.
-     */
-    private static final int ALT_PORT = 8282;
-    private static final Map<String, String> HEADERS = Map.ofEntries(entry("Authorization", "Bearer supersecret"));
-
-    //CHECKSTYLE:OFF
-    /*
-     * Checkstyle is disabled here because it wants opaContainer to 'be private and have accessor methods', which seems
-     * pointless and will probably mess up test containers.
-     */
-    @Container
-    public static GenericContainer<?> opaContainer = new GenericContainer<>(
-        new ImageFromDockerfile()
-            // .withFileFromClasspath(path_in_build_context, path_in_resources_dir)
-            .withFileFromClasspath("Dockerfile", "opa.Dockerfile")
-            .withFileFromClasspath("nginx.conf", "nginx.conf")
-            .withFileFromClasspath("entrypoint.sh", "entrypoint.sh")
-    )
-        .withExposedPorts(OPA_PORT, ALT_PORT)
-        .withFileSystemBind("./testdata/simple", "/policy", BindMode.READ_ONLY)
-        .withCommand("run -s --authentication=token --authorization=basic --bundle /policy --addr=0.0.0.0:" + OPA_PORT)
-        .withLogConsumer(new Slf4jLogConsumer(LoggerFactory.getLogger(OPAAuthorizationManagerTest.class)));
-    //CHECKSTYLE:ON
-
-    @Mock
-    private Supplier<Authentication> authenticationSupplier;
-
-    @Mock
-    private RequestAuthorizationContext context;
-
-    @Mock
-    private HttpServletRequest httpServletRequest;
-
-    private final ObjectMapper mapper = new ObjectMapper();
-
-    private String address;
-    private String altAddress;
-
-    @DynamicPropertySource
-    static void registerOpaProperties(DynamicPropertyRegistry registry) {
-        registry.add("opa.url",
-            () -> String.format("http://%s:%d", opaContainer.getHost(), opaContainer.getMappedPort(OPA_PORT)));
-    }
-
-    @BeforeEach
-    public void setUp() {
-        MockitoAnnotations.openMocks(this);
-        when(context.getRequest()).thenReturn(httpServletRequest);
-
-        HashMap<String, String> mockHeaders = new HashMap<>();
-        mockHeaders.put("UnitTestHeader", "123abc");
-        when(httpServletRequest.getHeaderNames()).thenReturn(Collections.enumeration(mockHeaders.keySet()));
-        when(httpServletRequest.getHeader(anyString())).thenAnswer(invocation -> {
-            String headerName = invocation.getArgument(0);
-            return mockHeaders.get(headerName);
-        });
-
-        when(httpServletRequest.getServletPath()).thenReturn("unit/test");
-        when(httpServletRequest.getMethod()).thenReturn("GET");
-        when(httpServletRequest.getProtocol()).thenReturn("HTTP/1.1");
-        when(httpServletRequest.getRemoteHost()).thenReturn("example.com");
-        when(httpServletRequest.getRemoteAddr()).thenReturn("192.0.2.123");
-
-        address = "http://" + opaContainer.getHost() + ":" + opaContainer.getMappedPort(OPA_PORT);
-        altAddress = "http://" + opaContainer.getHost() + ":" + opaContainer.getMappedPort(ALT_PORT) + "/customprefix";
-    }
-
-    @AfterEach
-    public void dumpLogs() {
-        System.out.println("==== container logs from OPA container ====");
-        System.out.println(opaContainer.getLogs());
-    }
+class OPAAuthorizationManagerTest extends BaseIntegrationTest {
 
     /**
      * This test just makes sure that we can reach the OPAClient health endpoint and that it returns the expected body.
@@ -421,155 +303,5 @@ class OPAAuthorizationManagerTest {
         assertEquals("echo rule always allows", actualResponse.getReasonForDecision("en"));
         assertEquals("other reason key", actualResponse.getReasonForDecision("other"));
         assertEquals("echo rule always allows", actualResponse.getReasonForDecision("nonexistant"));
-    }
-
-    private Authentication createMockAuthentication() {
-        Authentication mockAuth = mock(Authentication.class);
-        when(mockAuth.getPrincipal()).thenReturn("testuser");
-        when(mockAuth.getCredentials()).thenReturn("letmein");
-
-        WebAuthenticationDetails details = new WebAuthenticationDetails(httpServletRequest);
-        when(mockAuth.getDetails()).thenReturn(details);
-
-        GrantedAuthority authority1 = new SimpleGrantedAuthority("ROLE_USER");
-        GrantedAuthority authority2 = new SimpleGrantedAuthority("ROLE_ADMIN");
-        Collection<? extends GrantedAuthority> authorities = Arrays.asList(authority1, authority2);
-        doReturn(authorities).when(mockAuth).getAuthorities();
-
-        when(mockAuth.isAuthenticated()).thenReturn(true);
-
-        return mockAuth;
-    }
-
-    /**
-     * This is used to create a mock {@link Authentication} object where most of the fields are null, to resolve
-     * exceptions when optional fields are omitted.
-     */
-    private Authentication createNullMockAuthentication() {
-        Authentication mockAuth = mock(Authentication.class);
-        when(mockAuth.getPrincipal()).thenReturn("testuser");
-        when(mockAuth.getCredentials()).thenReturn(null);
-        when(mockAuth.getDetails()).thenReturn(null);
-
-        doReturn(null).when(mockAuth).getAuthorities();
-
-        when(mockAuth.isAuthenticated()).thenReturn(true);
-
-        return mockAuth;
-    }
-
-    /**
-     * Convert the value to JSON and then retrieve the value at the specified path.<br/>
-     * Note that this does stringify all JSON types, including null, so there could be some slight shadowing problems.
-     */
-    private String jsonGet(Object root, String path) {
-        JsonNode jsonRoot = this.mapper.valueToTree(root);
-        return jsonRoot.at(path).asText();
-    }
-
-    /**
-     * List all JSON paths found under the object.
-     */
-    private List<String> jsonList(Object root) {
-        JsonNode jsonRoot = this.mapper.valueToTree(root);
-        List<String> paths = new ArrayList<>();
-        jsonList(jsonRoot, "", paths);
-        return paths;
-    }
-
-    private void jsonList(JsonNode node, String currentPath, List<String> paths) {
-        if (node.isValueNode()) {
-            paths.add(currentPath);
-        } else if (node.isObject()) {
-            node.fields().forEachRemaining(entry -> {
-                String fieldName = entry.getKey();
-                JsonNode childNode = entry.getValue();
-                jsonList(childNode, currentPath + "/" + fieldName, paths);
-            });
-        } else if (node.isArray()) {
-            for (int i = 0; i < node.size(); i++) {
-                jsonList(node.get(i), currentPath + "/" + i, paths);
-            }
-        }
-    }
-
-    /**
-     * Create a human-readable list of differences between two objects. The possible paths are enumerated with
-     * {@link #jsonList(Object)}, and they are retrieved using {@link #jsonGet(Object, String)}. This does mean that
-     * all values are compared stringly.
-     */
-    private List<String> jsonDiff(Object rootA, Object rootB) {
-        List<String> pathsA = jsonList(rootA);
-        List<String> pathsB = jsonList(rootB);
-        Set<String> pathSet = new HashSet<>(pathsA);
-        pathSet.addAll(pathsB);
-        List<String> paths = new ArrayList<>(pathSet);
-
-        List<String> results = new ArrayList<>();
-
-        for (int i = 0; i < paths.size(); i++) {
-            String valA = jsonGet(rootA, paths.get(i));
-            String valB = jsonGet(rootB, paths.get(i));
-            if (!Objects.equals(valA, valB)) {
-                results.add(String.format("%s: %s =/= %s", paths.get(i), valA, valB));
-            }
-        }
-
-        return results;
-    }
-
-    private String jsonPretty(Object root) {
-        try {
-            return mapper.writerWithDefaultPrettyPrinter().writeValueAsString(root);
-            //CHECKSTYLE:OFF
-        } catch (Exception e) {
-            //CHECKSTYLE:ON
-            return String.format("failed to pretty print JSON: %s", e);
-        }
-    }
-
-    @Nested
-    @Import(OPAPathSelectorTest.CustomOPAConfig.class)
-    public class OPAPathSelectorTest {
-
-        @Autowired
-        private OPAAuthorizationManager opaAuthorizationManager;
-
-        @Test
-        public void testOPAPathSelectorAlwaysTruePolicy() {
-            Authentication mockAuth = createMockAuthentication();
-            when(mockAuth.getPrincipal()).thenReturn("testuser_allowed");
-            when(authenticationSupplier.get()).thenReturn(mockAuth);
-            assertDoesNotThrow(() -> opaAuthorizationManager.verify(authenticationSupplier, context));
-        }
-
-        @Test
-        public void testOPAPathSelectorAlwaysFalsePolicy() {
-            Authentication mockAuth = createMockAuthentication();
-            when(mockAuth.getPrincipal()).thenReturn("testuser_denied");
-            when(authenticationSupplier.get()).thenReturn(mockAuth);
-            assertThrows(AccessDeniedException.class,
-                () -> opaAuthorizationManager.verify(authenticationSupplier, context));
-        }
-
-        @Order(Ordered.HIGHEST_PRECEDENCE)
-        @TestConfiguration
-        public static class CustomOPAConfig {
-            @Bean
-            public OPAClient opaClient(OPAProperties opaProperties) {
-                return new OPAClient(opaProperties.getUrl(), HEADERS);
-            }
-
-            @Bean
-            public OPAPathSelector opaPathSelector() {
-                return (authentication, requestAuthorizationContext, opaInput) -> {
-                    if (authentication.getPrincipal().equals("testuser_allowed")) {
-                        return "policy/decision_always_true";
-                    } else {
-                        return "policy/decision_always_false";
-                    }
-                };
-            }
-        }
     }
 }
